clean up scripts
