+++ servers/visualization/app.py
@@ -1,137 +1,15 @@
 import os
-import re
 import sys
-from umap import UMAP
-from hdbscan import HDBSCAN
-from sentence_transformers import SentenceTransformer
-from sklearn.feature_extraction.text import CountVectorizer
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import StandardScaler
-from scipy.cluster.hierarchy import linkage
-from scipy.cluster.hierarchy import ward, fcluster
-from bertopic import BERTopic
-from bertopic.representation import KeyBERTInspired, MaximalMarginalRelevance, OpenAI
-from bertopic.vectorizers import ClassTfidfTransformer
-from tenacity import retry, stop_after_attempt, wait_exponential
-import yaml
 import openai
-import logging
-import json
-import nltk
-import time
-import spacy
 import numpy as np
-from nltk.corpus import stopwords
-from nltk.tokenize import word_tokenize
-from bertopic import BERTopic
-from sentence_transformers import SentenceTransformer
-import plotly.express as px
 import plotly.graph_objects as go
-from jupyter_dash import JupyterDash
 import dash
 from dash import dcc, html
 from dash.dependencies import Input, Output, State
 import plotly.graph_objects as go
 import dash_bootstrap_components as dbc
-import plotly.express as px
 
-sys.path.append(os.path.abspath("genai-interactive-topics/code/"))
-
-from process_text import *
-from generate_description import *
-
-def load_bertopic_model(model_dir):
-    """Load the BERTopic model from the specified directory."""
-    model_path = os.path.join(model_dir, 'topic_model.pkl')
-    if not os.path.exists(model_path):
-        raise FileNotFoundError(f"Model file not found at: {model_path}")
-    return BERTopic.load(model_path)
-
-def load_json_files(model_dir):
-    """Load topic keywords and sizes from JSON files."""
-    keywords_path = os.path.join(model_dir, 'topic_keywords_with_labels_gpt.json')
-    sizes_path = os.path.join(model_dir, 'topic_sizes_with_labels_gpt.json')
-    if not os.path.exists(keywords_path):
-        raise FileNotFoundError(f"Keywords file not found at: {keywords_path}")
-    if not os.path.exists(sizes_path):
-        raise FileNotFoundError(f"Sizes file not found at: {sizes_path}")
-    with open(keywords_path, 'r') as f:
-        topic_keywords_with_labels = json.load(f)
-    with open(sizes_path, 'r') as f:
-        topic_sizes = json.load(f)
-    print(topic_sizes)
-    print("Number of topics:", len(topic_sizes))
-    print("Total number of chunks:", sum(topic_sizes.values()))
-    return topic_keywords_with_labels, topic_sizes
-    
-def load_cluster_descriptions(model_dir):
-    des_path = os.path.join(model_dir, 'cluster_descriptions_gpt.json')
-    if not os.path.exists(des_path):
-        raise FileNotFoundError(f"Keywords file not found at: {des_path}")
-    with open(des_path, 'r') as f:
-        cluster_descriptions = json.load(f)
-    return cluster_descriptions
-
-# Function to save cluster_descriptions into a JSON file
-def save_cluster_descriptions_to_json(model_dir, cluster_descriptions, output_filename="cluster_descriptions.json"):
-    # Convert keys to standard int type
-    cluster_descriptions = {int(key): value for key, value in cluster_descriptions.items()}
-    # Save to JSON
-    with open(os.path.join(model_dir, output_filename), "w", encoding="utf-8") as file:
-        json.dump(cluster_descriptions, file, ensure_ascii=False, indent=4)
-    print(f"Cluster descriptions saved to {output_filename}")
-
-# Function to cluster topics based on intertopic distances
-def cluster_topics(topic_model, coords, random_state=42):
-    # Set random seed for reproducibility
-    np.random.seed(random_state)
-    x_vals = coords['x']
-    y_vals = coords['y']
-
-    # Standardize the coordinates
-    coordinates = np.vstack((x_vals, y_vals)).T
-    scaler = StandardScaler()
-    scaled_coords = scaler.fit_transform(coordinates)
-
-    # Perform hierarchical clustering
-    linkage_matrix = linkage(scaled_coords, method='ward')
-    clusters = fcluster(linkage_matrix, t=1, criterion='distance')
-    return clusters
-
-# Function to order clusters by importance
-def order_clusters_by_importance(topic_model, clusters, coords, random_state=42):
-    # Set random seed for reproducibility
-    np.random.seed(random_state)
-    x_vals = coords['x']
-    y_vals = coords['y']
-    coordinates = np.vstack((x_vals, y_vals)).T
-
-    # Standardize the coordinates
-    scaler = StandardScaler()
-    scaled_coords = scaler.fit_transform(coordinates)
-
-    # Apply PCA to extract variance contributions
-    pca = PCA(n_components=2, random_state=random_state)  # Use 2 components to match x, y dimensions
-    pca_coords = pca.fit_transform(scaled_coords)
-    explained_variance = pca.explained_variance_ratio_
-    
-    # Assign PCA contributions to clusters
-    cluster_variance = {}
-    for cluster_id in np.unique(clusters):
-        cluster_points = pca_coords[np.array(clusters) == cluster_id]
-        # Sum variance contributions of all points in the cluster
-        cluster_variance[cluster_id] = np.sum(np.var(cluster_points, axis=0) * explained_variance)
-    # Order clusters by variance contribution
-    ordered_clusters = sorted(cluster_variance.items(), key=lambda x: x[1], reverse=True)
-    return ordered_clusters
-
-# Function to display top keywords for the selected topic
-def display_keywords(topic_label, topic_keywords_with_labels):
-    if topic_label in topic_keywords_with_labels:
-        keywords = topic_keywords_with_labels[topic_label]
-        return [html.H4(f"Top Keywords for {topic_label}", className="card-title"),
-                html.P(f"{', '.join(keywords)}", className="card-text", style={"font-size": "16px"})]
-    return html.P("No keywords available.", style={"font-size": "16px", "color": "black"})
+from utils import *
 
 def main(directory, rerun=False, key=None):
     # Load the BERTopic model and files   
@@ -145,24 +23,7 @@ def main(directory, rerun=False, key=None):
     label_to_id = {label: topic_id for topic_id, label in enumerate(topic_keywords_with_labels.keys())}
     topic_labels = list(topic_keywords_with_labels.keys())
 
-    # Generate Clustering 
-    clusters = cluster_topics(topic_model, coords, random_state=42)
-    ordered_clusters = order_clusters_by_importance(topic_model, clusters, coords, random_state=42)
-    # Create a mapping from the original cluster IDs to new ordered IDs
-    cluster_reorder_map = {old_id: new_id for new_id, (old_id, _) in enumerate(ordered_clusters, start=1)}
-    # Apply the mapping to reorder the cluster labels
-    reordered_clusters = np.array([cluster_reorder_map[c] for c in clusters])
-
-    if rerun == True:
-        if openai.api_key != None:
-            cluster_descriptions = generate_cluster_descriptions_GPT(reordered_clusters, topic_keywords_with_labels)
-            # Call the function to save the descriptions
-            save_cluster_descriptions_to_json(model_dir, cluster_descriptions, "cluster_descriptions_GPT.json")
-        else:
-            cluster_descriptions = generate_cluster_descriptions(reordered_clusters, topic_keywords_with_labels)
-            save_cluster_descriptions_to_json(model_dir, cluster_descriptions, "cluster_descriptions.json")
-    else:
-        cluster_descriptions = load_cluster_descriptions(model_dir)
+    cluster_descriptions = load_cluster_descriptions(model_dir)
     
     # Initialize the Dash app with a Bootstrap theme
     app = dash.Dash(__name__, external_stylesheets=[dbc.themes.BOOTSTRAP])
@@ -196,7 +57,6 @@ def main(directory, rerun=False, key=None):
         scatter_trace = go.Scatter(
             x=x_vals,
             y=y_vals,
-            #mode='markers+text',
             mode='markers',
             text=topic_labels,
             hovertext=hover_texts,
@@ -206,21 +66,6 @@ def main(directory, rerun=False, key=None):
             textfont=dict(size=12, color='black'))
     
         fig = go.Figure(data=[scatter_trace])
-        #fig.update_layout(
-        #    title="Intertopic Distance Map",
-        #    height=700,
-        #    xaxis_title="PC1",
-        #    yaxis_title="PC2",
-        #    plot_bgcolor='white',
-        #    paper_bgcolor='white',
-        #    hovermode="closest",
-        #    font=dict(family="Arial", size=12, color="black"),
-        #    xaxis=dict(showgrid=True, gridcolor='lightgrey'),
-        #    yaxis=dict(showgrid=True, gridcolor='lightgrey'))
-        # Update layout to remove background and grid lines
-        # Update layout for a clean, square plot with axis lines
-        # Update layout for a square plot with axis lines but no grid or ticks
-        # Update layout to keep X and Y axis lines but remove grid and ticks
         fig.update_layout(
             title="Intertopic Distance Map",
             xaxis=dict(
@@ -246,8 +91,6 @@ def main(directory, rerun=False, key=None):
             plot_bgcolor="white",  # White background
             paper_bgcolor="white",  # White canvas
             autosize=True,
-            #width=700,  # Ensure square proportions
-            #height=700,  # Ensure square proportions
         )
         return fig
 
@@ -265,7 +108,6 @@ def main(directory, rerun=False, key=None):
                     title=f"Top {n_words} Keywords for Topic {topic_label}",
                     xaxis_title="Weight",
                     yaxis_title="Keywords",
-                    #height=500,
                     template="plotly_white")
                 return fig
         return go.Figure()
@@ -279,12 +121,11 @@ def main(directory, rerun=False, key=None):
                 html.P("Select a topic using the buttons or the dropdown below:", className="lead"),
                 html.Button('Previous Topic', id='prev-topic', n_clicks=0),
                 html.Button('Next Topic', id='next-topic', n_clicks=0),
-                html.Button('Clear Topic', id='clear-topic', n_clicks=0),
                 dcc.RadioItems(
                     id="topic-selector",
                     options=[{'label': label, 'value': label} for label in topic_keywords_with_labels.keys()],
                     labelStyle={'display': 'block'},
-                    style={"height": "300px", "overflow-y": "scroll"},
+                    style={"height": "300px", "overflowY": "scroll"},
                 ),
                 dbc.Button("Reset View", id="reset-button", color="primary", className="mt-3"),
             ], width=3),
@@ -300,7 +141,7 @@ def main(directory, rerun=False, key=None):
                     dbc.CardBody([
                         html.Div(id='keyword-display')
                     ])
-                ], style={"height": "100%", "background-color": "#e3f2fd"})
+                ], style={"height": "100%", "backgroundColor": "#e3f2fd"})
             ], width=6),
     
             dbc.Col([
@@ -311,7 +152,7 @@ def main(directory, rerun=False, key=None):
     
         dbc.Row([
             dbc.Col([
-                html.Div(id='cluster-descriptions', style={"margin-top": "20px", "padding": "10px", "background-color": "#f0f8ff", "border-radius": "10px"})
+                html.Div(id='cluster-descriptions', style={"marginTop": "20px", "padding": "10px", "backgroundColor": "#f0f8ff", "borderRadius": "10px"})
             ])
         ])
     ], fluid=True)
@@ -332,7 +173,7 @@ def main(directory, rerun=False, key=None):
         if triggered == 'reset-button.n_clicks':
             # Reset view and clear selections
             return create_intertopic_distance_plot(coords), \
-                   html.H5("Select a topic to see the top keywords.", style={"font-size": "18px", "color": "black"}), \
+                   html.H5("Select a topic to see the top keywords.", style={"fontSize": "18px", "color": "black"}), \
                    go.Figure()
     
         if triggered == 'intertopic-plot.clickData':
@@ -348,15 +189,8 @@ def main(directory, rerun=False, key=None):
             return intertopic_plot, keywords_display, term_bar_chart
     
         return create_intertopic_distance_plot(coords), \
-               html.H5("Select a topic to see the top keywords.", style={"font-size": "18px", "color": "black"}), \
+               html.H5("Select a topic to see the top keywords.", style={"fontSize": "18px", "color": "black"}), \
                go.Figure()
-
-    # Adjust overlapping labels
-    def adjust_label_positions(x_vals, y_vals, labels):
-        jitter_strength = 0.03
-        adjusted_x = x_vals + np.random.uniform(-jitter_strength, jitter_strength, len(x_vals))
-        adjusted_y = y_vals + np.random.uniform(-jitter_strength, jitter_strength, len(y_vals))
-        return adjusted_x, adjusted_y
 
     # Update cluster descriptions
     @app.callback(
@@ -370,14 +204,12 @@ def main(directory, rerun=False, key=None):
     # Navigation buttons to cycle through topics
     @app.callback(
         Output('topic-selector', 'value'),
-        [Input('prev-topic', 'n_clicks'), Input('next-topic', 'n_clicks'), Input('clear-topic', 'n_clicks')],
+        [Input('prev-topic', 'n_clicks'), Input('next-topic', 'n_clicks')],
         [State('topic-selector', 'value')]
     )
-    def navigate_topics(prev_clicks, next_clicks, clear_clicks, current_value):
+    def navigate_topics(prev_clicks, next_clicks, current_value):
         current_index = topic_labels.index(current_value) if current_value else 0
-        if clear_clicks > 0:
-            return None
-        elif prev_clicks > 0 and current_index > 0:
+        if prev_clicks > 0 and current_index > 0:
             return topic_labels[current_index - 1]
         elif next_clicks > 0 and current_index < len(topic_labels) - 1:
             return topic_labels[current_index + 1]
@@ -388,12 +220,8 @@ def main(directory, rerun=False, key=None):
             return None  # Reset topic selection to None
 
     # Run the app
-    app.run_server(debug=True, host="0.0.0.0", port=8092)
+    app.run_server(debug=True, port=8092)
 
 
 if __name__ == "__main__":
-    import argparse
-    parser = argparse.ArgumentParser()
-    parser.add_argument("--model_dir", type=str, required=True, help="Directory containing the trained BERTopic model and related JSON files")
-    args = parser.parse_args()
-    main(args.model_dir)
+    main("model_info/model_info_all")
