+++ servers/visualization/requirements.txt
@@ -1,0 +1,15 @@
+numpy==1.26.4
+scikit-learn==1.5.2
+openai==0.28.0
+hdbscan==0.8.33
+sentence-transformers==2.5.1
+bertopic==0.16.0
+plotly==5.20.0
+torch==2.2.1
+pandas==2.2.0
+spacy==3.7.4
+scipy==1.12.0
+tenacity==8.2.3
+dash==2.18.1
+dash-bootstrap-components==1.6.0
+flask>=2.0.0
